add navigate

+++ user/src/main/java/org/the_chance/honeymart/ui/feature/market/MarketAdapter.kt
@@ -10,5 +10,5 @@ class MarketAdapter(listener: MarketInteractionListener) : BaseAdapter<MarketUiS
 }
 
 interface MarketInteractionListener: BaseInteractionListener{
-    fun onClickMarket(id: Int)
+    fun onClickMarket(id: Long)
 }
